Put benefits ahead of requirements and fixed requirements
Previously still mostly listed hoops through which we expected builders to jump.

Have since added more benefits and made it clear that most requirements are soft ones.

The main thing is to build something useful.  You do NOT have to start it by filling out a form and using our repository.  That is NOT consistent with the life cycle of a CNCF Sandbox project for example.  Rather - projects start wherever and then can decide they want to be supported by the ODC.  We call that becoming a Sandbox project.